fixed C++ ambiguity
X-Trac-URL: https://ssl.icu-project.org/trac/ticket/2039



git-svn-id: http://source.icu-project.org/repos/icu/icu/trunk@12494 0a8b766c-62c9-45f9-954d-7f2943c9ead0

+++ source/test/intltest/itrbnf.cpp
@@ -219,7 +219,7 @@ IntlTestRBNF::TestAPI() {
   }
   result.remove();
   expected = "four";
-  formatter->format((long)4,result);
+  formatter->format((int)4,result);
   if(result != expected) {
       errln("Formatted 4, expected " + expected + " got " + result);
   } else {
